Fix libgphoto2_port compilation warnings.
* gphoto2-port-info-list.c:
  * gp_port_info_list_append(...): Variables 'generic' and 'i' are changed to unsigned types as they are used in trivial 'for' loops already in unsigned comparison mode.
  * gp_port_info_list_get_info(...): Cast variable 'n' to unsigned in two places to avoid signed-unsigned compare warnings. This does not change behavior.
* gphoto2-port-log.c:
  * gp_log_remove_func(...): Change variable 'i' to unsigned type as it is used in trivial 'for' loop in unsigned comparison mode. Also casts variable 'id' to unsigned to avoid signed-unsigned comparison. This does not change code behavior.

+++ libgphoto2_port/libgphoto2_port/gphoto2-port-info-list.c
@@ -176,7 +176,7 @@ gp_port_info_list_free (GPPortInfoList *list)
 int
 gp_port_info_list_append (GPPortInfoList *list, GPPortInfo info)
 {
-	int generic, i;
+	unsigned int generic, i;
 
 	C_PARAMS (list);
 
@@ -465,13 +465,13 @@ gp_port_info_list_get_info (GPPortInfoList *list, int n, GPPortInfo *info)
 
 	GP_LOG_D ("Getting info of entry %i (%i available)...", n, list->count);
 
-	C_PARAMS (n >= 0 && n < list->count);
+	C_PARAMS (n >= 0 && (unsigned int)n < list->count);
 
 	/* Ignore generic entries */
 	for (i = 0; i <= n; i++)
 		if (!strlen (list->info[i]->name)) {
 			n++;
-			C_PARAMS (n < list->count);
+			C_PARAMS ((unsigned int)n < list->count);
 		}
 
 	*info = list->info[n];
